Fixing
- fixing master book search

+++ src/main/java/com/test/library/main/dto/response/BookMasterDto.java
@@ -7,5 +7,5 @@ public record BookMasterDto(
     String title,
     String author,
     String synopsis,
-    Integer availableUnits
+    Long availableUnits
 ) { }
